pyproject.toml
[tool.poetry]
name = "mongars"
version = "0.7.1"
description = "Show unread emails in INBOX using Gnome Online Accounts"
authors = ["Chmouel Boudjnah <chmouel@chmouel.com>"]
keywords = ["email", "cli"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://github.com/chmouel/mongars"
repository = "https://github.com/chmouel/mongars"
include = ["LICENSE"]
classifiers = ["License :: OSI Approved :: Apache Software License"]

[tool.poetry.dependencies]
python = "^3.8"
PyGObject = "^3.42.0"
google-api-core = "^2.11.1"
google-auth = "^2.23.0"
google-auth-oauthlib = "^1.1.0"
google-api-python-client = "^2.99.0"

[tool.poetry.scripts]
mongars = "mongars.cli:main"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.12.2"
mypy = "^0.920"
yapf = "^0.31.0"
icecream = "^2.1.1"

[tool.ruff]
line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
